Default width/height resource limit is constrained by range of signed 32-bit integer. Coders verify that width/height/pixels limits have not been exceeded immediately, while still allowing 'ping' mode to work properly.

+++ magick/resource.c
@@ -91,7 +91,7 @@ static SemaphoreInfo
 /*
   Array must be in same order as ResourceType enum
 */
-#define PIXEL_LIMIT (LONG_MAX/(sizeof(PixelPacket)))
+#define PIXEL_LIMIT (INT_MAX/(sizeof(PixelPacket)))
 static ResourceInfo
   resource_info[] =
   {
